LPS-65351 No need for a SAQMetricConfig interface since there is only a single implementation (bean style)
