+++ modules/apps/foundation/portal-security/portal-security-service-access-quota-service/src/main/java/com/liferay/portal/security/service/access/quota/internal/SAQMetricConfig.java
@@ -17,10 +17,69 @@ package com.liferay.portal.security.service.access.quota.internal;
 /**
  * @author Stian Sigvartsen
  */
-public interface SAQMetricConfig {
+public class SAQMetricConfig {
 
-	public abstract String getMetricName();
+	public SAQMetricConfig(String metricName, String pattern) {
+		_metricName = metricName;
+		_pattern = pattern;
+	}
 
-	public abstract String getPattern();
+	@Override
+	public boolean equals(Object obj) {
+		if (!(obj instanceof SAQMetricConfig)) {
+			return false;
+		}
+
+		SAQMetricConfig config2 = (SAQMetricConfig)obj;
+
+		if (getMetricName() != null) {
+			if (!getMetricName().equals(config2.getMetricName())) {
+				return false;
+			}
+		}
+		else if (config2.getMetricName() != null) {
+			return false;
+		}
+
+		if (getPattern() != null) {
+			if (!getPattern().equals(config2.getPattern())) {
+				return false;
+			}
+		}
+		else if (config2.getPattern() != null) {
+			return false;
+		}
+
+		return true;
+	}
+
+	public String getMetricName() {
+		return _metricName;
+	}
+
+	public String getPattern() {
+		return _pattern;
+	}
+
+	@Override
+	public int hashCode() {
+		return toString().hashCode();
+	}
+
+	public void setName(String metricName) {
+		_metricName = metricName;
+	}
+
+	public void setPattern(String pattern) {
+		_pattern = pattern;
+	}
+
+	@Override
+	public String toString() {
+		return _metricName + (_pattern != null ? "=" + _pattern : "");
+	}
+
+	private String _metricName;
+	private String _pattern;
 
 }
